create filepicker input in ui
